Lógica Usuarios
1) Funcionalidad Mapa.
2) Correo por varios intentos.

+++ src/com/unbosque/dao/impl/MovimientoDAOImpl.java
@@ -26,7 +26,7 @@ public class MovimientoDAOImpl implements DaoGeneral {
 	}
 
 	@Override
-	public Object get(long id) {
+	public Object get(int id) {
 		try {			
 			Session session = HibernateUtil.getSessionFactory().openSession();
 			Movimiento object = (Movimiento) session.load(Movimiento.class, id);
